requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
dnspython==2.4.2